refactor: combines mrq and render preset retrieval into 1 collector
